+++ nuxt.config.ts
@@ -56,5 +56,5 @@ export default defineNuxtConfig({
       cookieKey: 'i18n_redirected',
       redirectOn: 'root'
     }
-  }
+  },
 })
