+++ pages/sign-up.vue
@@ -92,7 +92,7 @@ const checkedVisible = ref(false)
 
             <v-card-text class="text-subtitle-1 d-flex justify-start flex-grow-0">
                 <RouterLink
-                    :to="localePath('/login')"
+                    :to="localePath('/')"
                     class="text-blue text-decoration-none"
                 >
                     {{$t('sign.back')}}
